removed the batching from the script
it was slow i.e. the time to copy the files around was almost as slow as the processing itself

Better to just run the entire script in parallel on multiple videos. Concern now is that I think rembg -p will not run incrementally if there was a previous incomplete run